requirements.txt
flask>=2.0.0,<3.0.0
flask-sqlalchemy>=3.0.0,<4.0.0
flask-migrate>=3.1.0,<4.0.0
flask-mail>=0.9.1,<1.0.0
flask-bcrypt>=1.0.1,<2.0.0
selenium>=4.0.0,<5.0.0
webdriver-manager>=3.8.0,<4.0.0
sqlalchemy>=2.0.0,<3.0.0
pyjwt>=2.0.0,<3.0.0
flask-jwt-extended>=4.4.4,<5.0.0
werkzeug>=2.0.0,<3.0.0
pyotp>=2.6.0,<3.0.0
# python-alipay-sdk>=3.0.0,<4.0.0
qrcode>=7.0.0,<8.0.0
pillow>=9.0.0,<10.0.0; python_version<"3.13"
pillow>=10.0.0,<11.0.0; python_version>="3.13"
requests>=2.25.0,<3.0.0
apscheduler>=3.9.0,<4.0.0
schedule>=1.1.0,<2.0.0
gunicorn>=20.0.0,<21.0.0
